fixed bugs concerning non number entries where numbers are expected; fixed logic in workshop parser

+++ source_code/blocks/BlockSequence.hpp
@@ -72,7 +72,7 @@ public:
         return time_left;
     }
 
-    void set_time_left(int amount){
+    void set_time_left(double amount){
         time_left = amount;
     }
